WIP - begin testing possible fixes.
The repository has now been configured so that a custom patch file
can be applied when building the test app. However, the fix being
checked-in in this commit does not yet work (crashes when using the
AVX-512 code path). This fix will be experimented with further until
it works correctly.