Mention that the build folder needs to be removed.
vcpkg seems to not pick up on the vcpkg-configuration.json change
that points it to a different ports folder, which can result in
a rebuilt binary using the previously pointed to OpenSSL binaries.
As a result, the bug fix may not be applied or removed when changing
the overlay ports folder. Removing the build folder appears to be
sufficient to get vcpkg to detect the ports folder change.