Update project to create three executables.
The first two executables are the original MCVE executable, which dumps XMM
register state to the console so that the bug is directly visible via the
printed values. The second executable is the updated code that tests multiple
OpenSSL ciphers for the bug but does not print XMM registers' values. The
last executable is a WIP that aims to generate code that relies on the XMM
registers' state and will behave differently if the OpenSSL bug is triggered.